tetragon: Move argument helpers code to args.go
Moving argument helpers code to args.go to structure the code better,
because it will be used from uprobe code in following changes.
